requirements.txt
streamlit
plotly
google-cloud-bigquery